+++ packages/frontend/pages/home/index.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
 import React from "react";
-import Feed from "../../components/Feed";
+import Feed from "../../components/FeedPage/Feed";
 
 const Home: NextPage = () => {
